+++ lib/ui/wallet/WalletMealPlanPanel.dart
@@ -17,8 +17,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/service/FlexUI.dart';
+import 'package:illinois/ui/wallet/WalletHomePanel.dart';
 import 'package:illinois/ui/wallet/WalletIlliniCashPanel.dart';
-import 'package:rokwire_plugin/service/app_datetime.dart';
 import 'package:illinois/model/IlliniCash.dart';
 import 'package:rokwire_plugin/service/auth2.dart';
 import 'package:illinois/service/IlliniCash.dart';
@@ -79,26 +79,31 @@ class WalletMealPlanPanel extends StatelessWidget {
 /////////////////////////////////
 // WalletMealPlanContentWidget
 
-class WalletMealPlanContentWidget extends StatefulWidget {
+class WalletMealPlanContentWidget extends StatefulWidget with WalletHomeContentWidget {
 
   final double headerHeight;
   WalletMealPlanContentWidget({super.key, this.headerHeight = 0});
 
   @override
   _WalletMealPlanContentWidgetState createState() => _WalletMealPlanContentWidgetState();
+
+  @override
+  Color get backgroundColor => Styles().colors.fillColorPrimaryVariant;
 }
 
 class _WalletMealPlanContentWidgetState extends State<WalletMealPlanContentWidget> implements NotificationsListener {
   bool _authLoading = false;
 
   bool _illiniCashLoading = false;
+  bool _showHistory = false;
+
+  List<BaseTransaction>? _mealPlanTransactions;
   bool _mealPlanTransactionsLoading = false;
+
+  List<BaseTransaction>? _cafeCreditTransactions;
   bool _cafeCreditTransactionsLoading = false;
 
-  DateTime? _startDate;
-  DateTime? _endDate;
-  List<BaseTransaction>? _mealPlanTransactions;
-  List<BaseTransaction>? _cafeCreditTransactions;
+  final int _historyNumberOfDays = 14;
 
   _WalletMealPlanContentWidgetState();
 
@@ -110,8 +115,13 @@ class _WalletMealPlanContentWidgetState extends State<WalletMealPlanContentWidge
       IlliniCash.notifyEligibilityUpdated,
     ]);
 
-    _loadBallance();
-    _loadThisMonthHistory();
+    _illiniCashLoading = (IlliniCash().ballance == null);
+    IlliniCash().updateBalance().then((_) {
+      setStateIfMounted(() {
+        _illiniCashLoading = false;
+      });
+    });
+
     super.initState();
   }
 
@@ -121,89 +131,99 @@ class _WalletMealPlanContentWidgetState extends State<WalletMealPlanContentWidge
     super.dispose();
   }
 
+  // NotificationsListener
+
+  @override
+  void onNotification(String name, dynamic param) {
+    if (name == IlliniCash.notifyBallanceUpdated) {
+      setState(() {});
+      _updateHistory();
+    }
+    else if (name == IlliniCash.notifyEligibilityUpdated) {
+      setState(() {});
+    }
+    else if (name == IlliniCash.notifyPaymentSuccess) {
+      _updateHistory();
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     return _authLoading ? _buildLoading() : _buildBody();
   }
 
   Widget _buildBody() =>
-    Container(child:
+    Container(color: widget.backgroundColor, child:
       Column(crossAxisAlignment: CrossAxisAlignment.stretch,
         children: <Widget>[
           _buildContentHeader(),
           _buildSettingsHeader(Localization().getStringEx("panel.settings.meal_plan.heading.text", "University Housing Meal Plan"), 'dining'),
           _buildMealPlanSection(),
-          _buildMealPlanHistory(),
-          _buildCafeCreditHistory(),
-          _buildBalancePeriodViewPicker(),
+          _buildHistory(),
         ],
       ),
     );
 
-  Widget _buildMealPlanHistory(){
-    return Column(
-      children: <Widget>[
-        _buildSettingsHeader(Localization().getStringEx(
-            "panel.settings.meal_plan.classic_meal_history.text", "Classic Meal History"), 'calendar'),
-        _buildBalanceTableRow(_mealPlanTransactionsLoading, _mealPlanTransactions),
-        Container(height: 20,),
-      ],
+  Widget _buildHistory() =>
+    Padding(padding: const EdgeInsets.symmetric(vertical: 24), child:
+      Column(children: <Widget>[
+        Padding(padding: EdgeInsets.symmetric(horizontal: 16), child:
+          RoundedButton(
+            label: Localization().getStringEx('panel.settings.meal_plan.button.view_history.title', 'View History'),
+            hint: Localization().getStringEx('panel.settings.meal_plan.button.view_history.hint', ''),
+            textStyle: Styles().textStyles.getTextStyle("widget.button.title.enabled"),
+            backgroundColor: Colors.white,
+            borderColor: Styles().colors.fillColorSecondary,
+            onTap: _onTapViewHistory,
+          ),
+        ),
+        if (_showHistory)
+          ...<Widget>[
+            _buildMealPlanHistory(),
+            _buildCafeCreditHistory(),
+          ],
+      ],),
     );
-  }
 
-  Widget _buildCafeCreditHistory(){
-    return Column(
-      children: <Widget>[
-        _buildSettingsHeader(Localization().getStringEx(
-            "panel.settings.meal_plan.dining_dollars_history.text", "Dining Dollars History"), 'calendar'),
-        _buildBalanceTableRow(_cafeCreditTransactionsLoading, _cafeCreditTransactions),
-        Container(height: 20,),
-      ],
+  Widget _buildMealPlanHistory() =>
+    Padding(padding: const EdgeInsets.only(top: 24), child:
+      Column(children: <Widget>[
+        _buildSettingsHeader(Localization().getStringEx("panel.settings.meal_plan.classic_meal_history.text", "Classic Meal History"), 'calendar'),
+        _buildBalanceTableRow(_mealPlanTransactionsLoading, _mealPlanTransactions),
+      ],),
     );
-  }
+
+  Widget _buildCafeCreditHistory() =>
+    Padding(padding: const EdgeInsets.only(top: 24), child:
+      Column(children: <Widget>[
+        _buildSettingsHeader(Localization().getStringEx("panel.settings.meal_plan.dining_dollars_history.text", "Dining Dollars History"), 'calendar'),
+        _buildBalanceTableRow(_cafeCreditTransactionsLoading, _cafeCreditTransactions),
+      ],),
+    );
 
   Widget _buildContentHeader() =>
     Container(height: widget.headerHeight, color: Styles().colors.fillColorPrimaryVariant,);
 
-  Widget _buildSettingsHeader(String? title, String iconKey){
-    if (!Auth2().isLoggedIn) {
-      return Container();
-    }
-    return Semantics(
-      label: title,
-      header: true,
-      excludeSemantics: true,
-      child: Container(
-        color: Styles().colors.fillColorPrimaryVariant,
-//        height: 56,
-        child: Align(
-          alignment: Alignment.centerLeft,
-          child: Padding(
-            padding: EdgeInsets.symmetric(horizontal: 16),
-            child: Row(
-              children: <Widget>[
-                Styles().images.getImage(StringUtils.ensureNotEmpty(
-                    iconKey, defaultValue: 'settings')) ?? Container(),
-                Expanded(child:
-                  Padding(
-                    padding: EdgeInsets.only(left: 12),
-                    child: Text(
-                      StringUtils.ensureNotEmpty(title),
-                      style: Styles().textStyles.getTextStyle("widget.heading.large"),
-                    ),
-                  )
+  Widget _buildSettingsHeader(String? title, String iconKey) =>
+    Semantics(label: title, header: true, excludeSemantics: true, child:
+      Container(color: Styles().colors.fillColorPrimaryVariant, child:
+        Align(alignment: Alignment.centerLeft, child:
+          Padding(padding: EdgeInsets.symmetric(horizontal: 16), child:
+            Row(children: <Widget>[
+              Styles().images.getImage(StringUtils.ensureNotEmpty(iconKey, defaultValue: 'settings')) ?? Container(),
+              Expanded(child:
+                Padding(padding: EdgeInsets.only(left: 12), child:
+                  Text(StringUtils.ensureNotEmpty(title), style: Styles().textStyles.getTextStyle("widget.heading.large"),),
                 )
-              ],
-            ),
+              )
+            ],),
           ),
         ),
       ),
     );
-  }
 
   Widget _buildMealPlanSection() {
     List<Widget> widgets = [];
-    widgets.add(Padding(padding: EdgeInsets.only(top: 16)));
 
     if (Auth2().isLoggedIn) {
       if (_illiniCashLoading) {
@@ -244,7 +264,7 @@ class _WalletMealPlanContentWidgetState extends State<WalletMealPlanContentWidge
       }
 
     }
-    else if (_canSignIn) {
+    else if (FlexUI().isAuthenticationAvailable) {
       widgets.add(Padding(
         padding: EdgeInsets.only(left: 20, right: 20, bottom: 16),
         child: RoundedButton(
@@ -265,113 +285,77 @@ class _WalletMealPlanContentWidgetState extends State<WalletMealPlanContentWidge
       widgets.add(_buildPrivacyAlertMessage());
     }
 
-    return Stack(
-      alignment: Alignment.center,
-      children: <Widget>[
-        Container(
-            color: Colors.white,
-            child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: widgets)
+    return Padding(padding: EdgeInsets.only(top: 8), child:
+      Stack(children: <Widget>[
+        Container(color: Colors.white, child:
+          Column(crossAxisAlignment: CrossAxisAlignment.start, children: widgets)
         ),
-        _illiniCashLoading ? CircularProgressIndicator() : Container()
-      ],
+        if (_illiniCashLoading)
+          Positioned.fill(child:
+            Center(child:
+              Padding(padding: EdgeInsets.symmetric(vertical: 24), child:
+                SizedBox(width: 24, height: 24, child:
+                  CircularProgressIndicator(color: Styles().colors.fillColorSecondary, strokeWidth: 3,)
+                )
+              ),
+            ),
+          ),
+      ],),
     );
   }
 
-  Widget _buildBalancePeriodViewPicker() {
-    if (!Auth2().isLoggedIn) {
-      return Container();
-    }
-    return Padding(
-      padding: EdgeInsets.only(left: 16, right: 16, top: 20, bottom: 24),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 10),
-            child: Text(
-                Localization().getStringEx("panel.settings.meal_plan.label.custom_period", "Custom Period"),
-                style: Styles().textStyles.getTextStyle("widget.detail.regular")
-            ),
-          ),
-          Padding(
-            padding: EdgeInsets.only(bottom: 16), child: Row(children: <Widget>[
-            _DateLabel(label: Localization().getStringEx('panel.settings.meal_plan.label.start_date', 'Start Date'),),
-            Container(width: 8,),
-            Expanded(child: _DateValue(title: StringUtils.ensureNotEmpty(
-                _getFormattedDate(_startDate)),
-              label: Localization().getStringEx('panel.settings.meal_plan.button.start_date.title', 'Start Date'),
-              hint: Localization().getStringEx('panel.settings.meal_plan.button.start_date.hint', ''),
-              onTap: _onStartDateTap,))
-          ],),),
-          Padding(
-            padding: EdgeInsets.only(bottom: 16), child: Row(children: <Widget>[
-            _DateLabel(label: Localization().getStringEx('panel.settings.meal_plan.label.end_date', 'End Date'),),
-            Container(width: 8,),
-            Expanded(child: _DateValue(
-              title: StringUtils.ensureNotEmpty(
-                  _getFormattedDate(_endDate)),
-              label: Localization().getStringEx('panel.settings.meal_plan.button.end_date.title', 'End Date'),
-              hint: Localization().getStringEx('panel.settings.meal_plan.button.end_date.hint', ''),
-              onTap: _onEndDateTap,))
-          ],),),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly, children: <Widget>[
-            Expanded(child: RoundedButton(
-              label: Localization().getStringEx('panel.settings.meal_plan.button.view_history.title', 'View History'),
-              hint: Localization().getStringEx('panel.settings.meal_plan.button.view_history.hint', ''),
-              textStyle: Styles().textStyles.getTextStyle("widget.button.title.enabled"),
-              backgroundColor: Colors.white,
-              borderColor: Styles().colors.fillColorSecondary,
-              onTap: _onTapViewHistory,)),
-          ],)
-        ],),);
-  }
-
   Widget _buildBalanceTableRow(bool loadingFlag, List<BaseTransaction>? transactionList) {
-    if(Auth2().isLoggedIn) {
-      if (loadingFlag) {
-        return Center(child: Padding(padding: EdgeInsets.only(bottom: 20),
-          child: CircularProgressIndicator(),),);
-      }
-      if (_startDate == null || _endDate == null ||
-          _startDate!.isAfter(_endDate!)) {
-        String text = Localization().getStringEx(
-            'panel.settings.meal_plan.transactions.message.start_end_validation.text',
-            'Start date must be before end date');
-        return Semantics(
-          label: text, hint: Localization().getStringEx(
-            'panel.settings.meal_plan.transactions.message.start_end_validation.hint',
-            ''), excludeSemantics: true, child: Center(child: Padding(
-          padding: EdgeInsets.only(left: 40, right: 40, bottom: 20),
-          child: Text(text,
-            style: Styles().textStyles.getTextStyle("widget.message.regular.fat"),),),),);
-      }
-      if (transactionList == null || transactionList.isEmpty) {
-        String text = Localization().getStringEx(
-            'panel.settings.meal_plan.transactions.message.no_transactions.text',
-            'There is no transactions for the selected period');
-        return Semantics(label: text, hint: Localization().getStringEx(
-            'panel.settings.meal_plan.transactions.message.no_transactions.hint',
-            ''), excludeSemantics: true, child: Center(child: Padding(
-          padding: EdgeInsets.only(left: 40, right: 40, bottom: 20),
-          child: Text(text,
-            textAlign: TextAlign.center,
-            style: Styles().textStyles.getTextStyle("widget.message.regular.fat"),),),),);
-      }
-      String dateHeader = Localization().getStringEx(
-          'panel.settings.meal_plan.label.date', 'Date');
-      String locationHeader = Localization().getStringEx(
-          'panel.settings.meal_plan.label.location', 'Location');
-      String descriptionHeader = Localization().getStringEx(
-          'panel.settings.meal_plan.label.description', 'Description');
-      String amountHeader = Localization().getStringEx(
-          'panel.settings.meal_plan.label.amount', 'Amount');
+    if (loadingFlag) {
+      return Padding(padding: EdgeInsets.only(top: 8), child:
+        Container(color: Styles().colors.white, child:
+          Row(children: [
+            Expanded(child:
+              Center(child:
+                Padding(padding: EdgeInsets.symmetric(vertical: 24), child:
+                  SizedBox(width: 24, height: 24, child:
+                    CircularProgressIndicator(color: Styles().colors.fillColorSecondary, strokeWidth: 3,)
+                  )
+                ),
+              )
+            )
+          ],),
+        ),
+      );
+    }
+    else if (CollectionUtils.isEmpty(transactionList)) {
+      String text = (transactionList != null) ?
+        Localization().getStringEx('panel.settings.meal_plan.transactions.message.no_transactions.text', 'There are no transactions in the last {{number_of_days}} days.').replaceAll('{{number_of_days}}', _historyNumberOfDays.toString()) :
+        Localization().getStringEx('panel.settings.meal_plan.transactions.message.failed_transactions.text', 'Failed to load transactions.');
+      String hint = (transactionList != null) ?
+        Localization().getStringEx('panel.settings.meal_plan.transactions.message.no_transactions.hint', '') :
+        Localization().getStringEx('panel.settings.meal_plan.transactions.message.failed_transactions.hint', '');
+      return Padding(padding: EdgeInsets.only(top: 8), child:
+        Semantics(label: text, hint: hint, excludeSemantics: true, child:
+          Container(color: Styles().colors.white, child:
+            Row(children: [
+              Expanded(child:
+                Center(child:
+                  Padding(padding: EdgeInsets.only(left: 32, right: 32, top: 16, bottom: 16), child:
+                    Text(text, textAlign: TextAlign.center, style: Styles().textStyles.getTextStyle("widget.message.regular.fat"),),
+                  ),
+                ),
+              ),
+            ]),
+          ),
+        ),
+      );
+    }
+    else {
+      String dateHeader = Localization().getStringEx('panel.settings.meal_plan.label.date', 'Date');
+      String locationHeader = Localization().getStringEx('panel.settings.meal_plan.label.location', 'Location');
+      String descriptionHeader = Localization().getStringEx('panel.settings.meal_plan.label.description', 'Description');
+      String amountHeader = Localization().getStringEx('panel.settings.meal_plan.label.amount', 'Amount');
+
       List<Widget> dateWidgets =  [];
       List<Widget> locationWidgets =  [];
       List<Widget> descriptionWidgets =  [];
       List<Widget> amountViewWidgets =  [];
+
       //Headers
       dateWidgets.add(_buildBalanceTableHeaderItem(dateHeader));
       locationWidgets.add(_buildBalanceTableHeaderItem(locationHeader));
@@ -384,23 +368,23 @@ class _WalletMealPlanContentWidgetState extends State<WalletMealPlanContentWidge
       int locationLenght = locationHeader.length;
       int descriptionLenght = descriptionHeader.length;
       int amountLenght = amountHeader.length;
-      transactionList.forEach((BaseTransaction? balance) {
+      transactionList?.forEach((BaseTransaction? balance) {
+
         //date
         String date = balance!.dateString!;
         dateLenght = dateLenght < date.length ? date.length : dateLenght;
+
         //location
         String location = balance.location!;
-        locationLenght =
-        locationLenght < location.length ? location.length : locationLenght;
+        locationLenght = locationLenght < location.length ? location.length : locationLenght;
+
         //description
         String description = balance.description!;
-        descriptionLenght = descriptionLenght < description.length
-            ? description.length
-            : descriptionLenght;
+        descriptionLenght = descriptionLenght < description.length ? description.length : descriptionLenght;
+
         //balance
         String amount = balance.amount!;
-        amountLenght =
-        amountLenght < amount.length ? amount.length : amountLenght;
+        amountLenght =amountLenght < amount.length ? amount.length : amountLenght;
 
         dateWidgets.add(_buildBalanceTableItem(text: date));
         locationWidgets.add(_buildBalanceTableItem(text: location));
@@ -408,94 +392,65 @@ class _WalletMealPlanContentWidgetState extends State<WalletMealPlanContentWidge
         amountViewWidgets.add(_buildAmountView(amount));
       });
 
-      return Row(
-          children: [
-            Expanded(child:
-            SingleChildScrollView(
-              scrollDirection: Axis.horizontal,
-              child:
-              Container(
-                child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Container(
-                        width: (dateLenght * textSize) + 20,
-                        child:
-                        Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: dateWidgets),),
-                      Container(
-                          width: locationLenght * textSize + 20,
-                          child:
-                          Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: locationWidgets)),
-                      Container(
-                          width: descriptionLenght * textSize +
-                              16 /*the padding*/,
-                          child:
-                          Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: descriptionWidgets)),
-                    ]),
-              ),
-            )),
-            Container(
-                width: amountLenght * textSize + 16 /*the padding*/,
-                child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: amountViewWidgets)),
-          ]);
-    }
-    else{
-      return Container();
-    }
-  }
-
-  Widget _buildBalanceTableHeaderItem(String text){
-    return _buildBalanceTableItem(text: text, backColor: Styles().colors.fillColorPrimaryVariant,
-        showBorder: false,
-        textStyle: Styles().textStyles.getTextStyle("widget.heading.small.fat"));
-  }
-
-  Widget _buildBalanceTableItem({required String text, bool showBorder = true, Color? backColor, TextStyle? textStyle, TextAlign textAlign = TextAlign.left}) {
-    return
-      Container(
-          width: double.infinity,
-          height: 40,
-          alignment: Alignment.centerLeft,
-          decoration: BoxDecoration(
-              color: backColor ?? Styles().colors.background,
-              border: showBorder?
-              Border.all(
-                  color: Styles().colors.surfaceAccent,
-                  width: 1,
-                  style: BorderStyle.solid) :
-              Border.all(color: backColor ?? Styles().colors.background, width: 0)
-          ),
-          child: Padding(
-            padding: EdgeInsets.all(8),
-            child:
-            Row(
-              children: <Widget>[
-                Expanded(
-                  child: Text(text,
-                    maxLines: 1,
-                    textAlign: textAlign,
-                    style: textStyle!=null? textStyle: Styles().textStyles.getTextStyle("widget.item.small.thin"),),
+      return Row(children: [
+        Expanded(child:
+          SingleChildScrollView(scrollDirection: Axis.horizontal, child:
+            Container(child:
+              Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
+                Container(width: (dateLenght * textSize) + 20, child:
+                  Column(crossAxisAlignment: CrossAxisAlignment.start, children: dateWidgets),
                 ),
-              ],
+                Container(width: locationLenght * textSize + 20, child:
+                  Column(crossAxisAlignment: CrossAxisAlignment.start, children: locationWidgets)
+                ),
+                Container(width: descriptionLenght * textSize + 16 /*the padding*/, child:
+                  Column(crossAxisAlignment: CrossAxisAlignment.start, children: descriptionWidgets)
+                ),
+              ]),
             ),
-
           )
-      );
+        ),
+        Container(width: amountLenght * textSize + 16 /*the padding*/, child:
+          Column(crossAxisAlignment: CrossAxisAlignment.start, children: amountViewWidgets)
+        ),
+      ]);
+    }
   }
 
-  Widget _buildAmountView(String balance){
-    return _buildBalanceTableItem(text: balance, backColor: Styles().colors.white,
-        textAlign: TextAlign.right,
-        textStyle: Styles().textStyles.getTextStyle("widget.item.small.fat"));
-  }
+  Widget _buildBalanceTableHeaderItem(String text) =>
+    _buildBalanceTableItem(
+        text: text,
+        backColor: Styles().colors.fillColorPrimaryVariant,
+        showBorder: false,
+        textStyle: Styles().textStyles.getTextStyle("widget.heading.small.fat")
+    );
+
+  Widget _buildBalanceTableItem({required String text, bool showBorder = true, Color? backColor, TextStyle? textStyle, TextAlign textAlign = TextAlign.left}) =>
+    Container(
+      width: double.infinity,
+      height: 40,
+      alignment: Alignment.centerLeft,
+      decoration: BoxDecoration(
+        color: backColor ?? Styles().colors.background,
+        border: showBorder ?
+          Border.all(color: Styles().colors.surfaceAccent, width: 1, style: BorderStyle.solid) :
+          Border.all(color: backColor ?? Styles().colors.background, width: 0)
+      ), child:
+        Padding(padding: EdgeInsets.all(8), child:
+          Row(children: <Widget>[
+            Expanded(child:
+              Text(text,maxLines: 1, textAlign: textAlign, style: textStyle ?? Styles().textStyles.getTextStyle("widget.item.small.thin"),),
+            ),
+          ],),
+        )
+      );
+
+  Widget _buildAmountView(String balance) =>
+    _buildBalanceTableItem(text: balance,
+      backColor: Styles().colors.white,
+      textAlign: TextAlign.right,
+      textStyle: Styles().textStyles.getTextStyle("widget.item.small.fat")
+    );
 
   Widget _buildPrivacyAlertMessage() {
     final String iconMacro = '{{privacy_level_icon}}';
@@ -504,15 +459,14 @@ class _WalletMealPlanContentWidgetState extends State<WalletMealPlanContentWidge
     String privacyMsgStart = (0 < iconMacroPosition) ? privacyMsg.substring(0, iconMacroPosition) : '';
     String privacyMsgEnd = ((0 < iconMacroPosition) && (iconMacroPosition < privacyMsg.length)) ? privacyMsg.substring(iconMacroPosition + iconMacro.length) : '';
 
-    return Container(
-      padding: EdgeInsets.only(left: 16, right: 16, bottom: 16),
-      child: RichText(text: TextSpan(
-        style: Styles().textStyles.getTextStyle("panel.settings.heading.title.large"),
-        children: [
-          TextSpan(text: privacyMsgStart),
-          WidgetSpan(alignment: PlaceholderAlignment.middle, child: _buildPrivacyLevelIcon()),
-          TextSpan(text: privacyMsgEnd)
-        ])));
+    return Container(padding: EdgeInsets.only(left: 16, right: 16, bottom: 16), child:
+      RichText(text: TextSpan(style: Styles().textStyles.getTextStyle("panel.settings.heading.title.large"), children: [
+        TextSpan(text: privacyMsgStart),
+        WidgetSpan(alignment: PlaceholderAlignment.middle, child: _buildPrivacyLevelIcon()),
+        TextSpan(text: privacyMsgEnd)
+      ])
+      )
+    );
   }
 
   Widget _buildPrivacyLevelIcon() {
@@ -530,111 +484,45 @@ class _WalletMealPlanContentWidgetState extends State<WalletMealPlanContentWidge
     ),
   );
 
-  // Data Access
-
-  void _loadMealPlanTransactions() {
-    _showMealPlanTransactionsProgress(true);
-    IlliniCash().loadMealPlanTransactionHistory(_startDate, _endDate).then((
-        transactions) => _onMealPlanTransactionsLoaded(transactions));
-  }
-
-  void _loadCafeCreditTransactions() {
-    _showMealPlanTransactionsProgress(true);
-    IlliniCash().loadCafeCreditTransactionHistory(_startDate, _endDate).then((
-        transactions) => _onCafeCreditTransactionsLoaded(transactions));
-  }
-
-  void _loadBallance() {
-    _illiniCashLoading = (IlliniCash().ballance == null);
-    IlliniCash().updateBalance().then((_){
-      setState(() {
-        _illiniCashLoading = false;
-      });
-    });
-  }
-
-  void _loadThisMonthHistory() {
-    Analytics().logSelect(target: "This Month");
-    DateTime now = DateTime.now();
-    DateTime lastMonth = now.subtract(Duration(
-      days: 30,
-    ));
-    _startDate = lastMonth;
-    _endDate = now;
-    _loadMealPlanTransactions();
-    _loadCafeCreditTransactions();
-  }
-
   // Handlers
-
-  void _onMealPlanTransactionsLoaded(List<MealPlanTransaction>? transactions) {
-    _showMealPlanTransactionsProgress(false, changeState: false);
-    if (mounted) {
-      setState(() {
-        _mealPlanTransactions = transactions;
-      });
-    }
-  }
-
-  void _onCafeCreditTransactionsLoaded(List<CafeCreditTransaction>? transactions) {
-    _showCafeCreditTransactionsProgress(false, changeState: false);
-    if (mounted) {
-      setState(() {
-        _cafeCreditTransactions = transactions;
-      });
-    }
-  }
 
   void _onTapViewHistory() {
     Analytics().logSelect(target: "View History");
-    _loadMealPlanTransactions();
-    _loadCafeCreditTransactions();
+    if (_showHistory == false) {
+      setState(() {
+        _showHistory = true;
+      });
+    }
+    _updateHistory();
   }
 
-  void _onStartDateTap() {
-    Analytics().logSelect(target: "Start date");
-    DateTime initialDate = _startDate ?? DateTime.now();
-    DateTime firstDate =
-    DateTime.fromMillisecondsSinceEpoch(initialDate.millisecondsSinceEpoch)
-        .add(Duration(days: -365));
-    DateTime lastDate =
-    DateTime.fromMillisecondsSinceEpoch(initialDate.millisecondsSinceEpoch)
-        .add(Duration(days: 365));
-    showDatePicker(
-      context: context,
-      firstDate: firstDate,
-      lastDate: lastDate,
-      initialDate: initialDate,
-      builder: (BuildContext context, Widget? child) {
-        return Theme(
-          data: ThemeData.light(),
-          child: child!,
-        );
-      },
-    ).then((selectedDateTime) => _onStartDateChanged(selectedDateTime));
-  }
+  void _updateHistory() {
+    DateTime endDate = DateTime.now();
+    DateTime startDate = DateTimeUtils.midnight(endDate.subtract(Duration(days: _historyNumberOfDays))) ?? endDate;
 
-  void _onEndDateTap() {
-    Analytics().logSelect(target: "End date");
-    DateTime initialDate = _endDate ?? DateTime.now();
-    DateTime firstDate =
-    DateTime.fromMillisecondsSinceEpoch(initialDate.millisecondsSinceEpoch)
-        .add(Duration(days: -365));
-    DateTime lastDate =
-    DateTime.fromMillisecondsSinceEpoch(initialDate.millisecondsSinceEpoch)
-        .add(Duration(days: 365));
-    showDatePicker(
-      context: context,
-      firstDate: firstDate,
-      lastDate: lastDate,
-      initialDate: initialDate,
-      builder: (BuildContext context, Widget? child) {
-        return Theme(
-          data: ThemeData.light(),
-          child: child!,
-        );
-      },
-    ).then((selectedDateTime) => _onEndDateChanged(selectedDateTime));
+    if (_mealPlanTransactionsLoading == false) {
+      setState(() {
+        _mealPlanTransactionsLoading = true;
+      });
+      IlliniCash().loadMealPlanTransactionHistory(startDate, endDate).then((List<MealPlanTransaction>? transactions) {
+        setStateIfMounted((){
+          _mealPlanTransactionsLoading = false;
+          _mealPlanTransactions = transactions;
+        });
+      });
+    }
+
+    if (_cafeCreditTransactionsLoading == false) {
+      setState(() {
+        _cafeCreditTransactionsLoading = true;
+      });
+      IlliniCash().loadCafeCreditTransactionHistory(startDate, endDate).then((List<CafeCreditTransaction>? transactions) {
+        setStateIfMounted((){
+          _cafeCreditTransactionsLoading = false;
+          _cafeCreditTransactions = transactions;
+        });
+      });
+    }
   }
 
   void _onTapLogIn() {
@@ -644,128 +532,11 @@ class _WalletMealPlanContentWidgetState extends State<WalletMealPlanContentWidge
       Auth2().authenticateWithOidc().then((Auth2OidcAuthenticateResult? result) {
         if (mounted) {
           setState(() { _authLoading = false; });
-          if (result == Auth2OidcAuthenticateResult.succeeded) {
-            _loadCafeCreditTransactions();
-            _loadMealPlanTransactions();
-          }
-          else if (result == Auth2OidcAuthenticateResult.failed) {
+          if (result == Auth2OidcAuthenticateResult.failed) {
             AppAlert.showDialogResult(context, Localization().getStringEx("logic.general.login_failed", "Unable to login. Please try again later."));
           }
         }
       });
     }
-  }
-
-  void _onStartDateChanged(DateTime? startDate) {
-    if(mounted) {
-      setState(() {
-        _startDate = startDate;
-      });
-    }
-  }
-
-  void _onEndDateChanged(DateTime? endDate) {
-    if(mounted) {
-      setState(() {
-        _endDate = endDate;
-      });
-    }
-  }
-
-  // Helpers
-
-  void _showMealPlanTransactionsProgress(bool loading, {bool changeState = true}) {
-    _mealPlanTransactionsLoading = loading;
-    if (changeState) {
-      if (mounted) {
-        setState(() {});
-      }
-    }
-  }
-
-  void _showCafeCreditTransactionsProgress(bool loading, {bool changeState = true}) {
-    _cafeCreditTransactionsLoading = loading;
-    if (changeState) {
-      if (mounted) {
-        setState(() {});
-      }
-    }
-  }
-
-  String? _getFormattedDate(DateTime? date) {
-    return AppDateTime().formatDateTime(
-        date, format: 'MM/dd/yyyy');
-  }
-
-  bool get _canSignIn{
-    return FlexUI().isAuthenticationAvailable;
-  }
-
-  // NotificationsListener
-
-  @override
-  void onNotification(String name, dynamic param) {
-    if (name == IlliniCash.notifyBallanceUpdated) {
-      setState(() {});
-      _loadCafeCreditTransactions();
-      _loadMealPlanTransactions();
-    }
-    else if (name == IlliniCash.notifyEligibilityUpdated) {
-      setState(() {
-
-      });
-    }
-    else if (name == IlliniCash.notifyPaymentSuccess) {
-      _loadCafeCreditTransactions();
-      _loadMealPlanTransactions();
-    }
-  }
-}
-
-/////////////////////////////////
-// _DateLabel
-
-class _DateLabel extends StatelessWidget {
-  final String? label;
-
-  _DateLabel({this.label});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(width: 70,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
-        Text(label!, style: Styles().textStyles.getTextStyle("widget.item.regular.thin"),),
-        Container(height: 2, color: Styles().colors.surfaceAccent,)
-      ],),);
-  }
-}
-
-
-/////////////////////////////////
-// _DateValue
-
-class _DateValue extends StatelessWidget {
-  final String? title;
-  final String? label;
-  final String? hint;
-  final GestureTapCallback? onTap;
-
-  _DateValue({this.title, this.label, this.hint, this.onTap});
-
-  @override
-  Widget build(BuildContext context) {
-    return Semantics(label: (0 < title!.length) ? title : label, hint: hint, button: true, excludeSemantics: true, child:InkWell(onTap: onTap, child: Column(children: <Widget>[
-      Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: <Widget>[
-          Expanded(child:
-            Text(title!,
-              overflow: TextOverflow.ellipsis,
-              style: Styles().textStyles.getTextStyle("widget.item.regular.fat"),),
-          ),
-          Styles().images.getImage('chevron-down', excludeFromSemantics: true) ?? Container(),
-        ],), Container(height: 2, color: Styles().colors.fillColorSecondary,)
-    ],),),);
   }
 }